requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
kafka-python>=2.0.2
pyspark>=3.3.0
PyYAML>=6.0
optuna>=3.0.0
plotly>=5.0.0
joblib>=1.1.0
scipy>=1.7.0
jupyter>=1.0.0